Append a risk-of-bias traffic-light plot to a forest plot

A wrapper for metafor::forest function, which adds a risk of bias traffic-light plot to the right-hand side of the forest plot. The heavy lifting for this function is done by metafor. Note that if not specified as additional arguments, this functions sets the header argument of metafor::forest() to TRUE.

rob_append_to_forest(
  res,
  rob_data,
  rob_tool = "ROB2",
  rob_colour = "cochrane",
  rob_psize = 2,
  rob_caption = TRUE,
  rob_legend = TRUE,
  ...
)

Arguments

res
Output from metafor meta-analysis function

rob_data
Risk of bias dataset in standard robvis format

rob_tool
The risk-of-bias assessment tool used to perform the assessments

rob_colour
Risk of bias colour scheme to use for the traffic-light plot

rob_psize
Size of the points in the appended risk of bias table

rob_caption
Logical specifying whether a message containing the risk-of-bias domains should be printed. Default is TRUE

rob_legend
Logical specifying whether a legend for the risk-of-bias plot should be shown. Default is TRUE.

...
Additional arguments to be passed to the metafor::forest() function

See also