Append a risk-of-bias traffic-light plot to a forest plot

A wrapper for metafor::forest function, which adds a risk of bias traffic-light plot to the right-hand side of the forest plot. The heavy lifting for this function is done by metafor. Note that if not specified as additional arguments, this functions sets the header argument of metafor::forest() to TRUE.

rob_forest(
  res,
  rob_tool = "ROB2",
  rob_me = NULL,
  rob_levels = NULL,
  title = NULL,
  rob_legend = TRUE,
  rob_legend_cex = 0.9,
  ...
)

Arguments

res

Output from metafor meta-analysis function

rob_tool

The risk-of-bias assessment tool used to perform the assessments

rob_me

Optional value defining the result of the Risk-Of-Bias due to Missing Evidence (ROB-ME) assessment for this synthesis. By default (rob_me = NULL), this is omitted from the plot.

rob_levels

Vector of judgments [e.g. c("Low","Some concerns","High","Critical")] that controls the ordering of subgroups within the plot

title

Text to use for plot title

rob_legend

Logical specifying whether a legend for the risk-of-bias plot should be shown. Default is TRUE.

rob_legend_cex

Expansion factor for the risk-of-bias legend

...

Additional arguments to be passed to the metafor::forest() function

See also