Save risk-of-bias plots to a file using sensible parameters

Save risk-of-bias plots to a file using sensible parameters

rob_save(
  rob_object,
  file = "rob_figure.png",
  height = "default",
  width = "default",
  dpi = 800
)

Arguments

rob_object
Object created using either rob_summary() or rob_traffic_light()

file
Destination file, with extension (e.g. "rob_figure.png")

height
Height of resulting image, in inches. Defaults to "default" which uses recommended values based on the number of studies included.

width
Width of resulting image, in inches. Defaults to "default" which uses recommended values based on the number of characters in the Study and Domain names. dpi

dpi
Plot resolution.

See also