Create "realistic" dummy risk of bias assessment data

This function returns N example risk of bias assessments for the tool specified, where N is set by the user. Assessments are "realistic" in that the judgment in the overall column is set to the highest judgement across the domains for a given study, reflecting the recommendations of the tool creators.

rob_dummy(n, tool = "ROB2", study = TRUE)

Arguments

n

Number of assessments to create

tool

Tool used for assessment (see rob_tools()). Default is "ROB2".

study

Should the returned dataframe contain a "Study" column. Default is true.

See also